% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod.R
\name{print.ecod}
\alias{print.ecod}
\title{Print Method for ECOD Objects}
\usage{
\method{print}{ecod}(x, ...)
}
\arguments{
\item{x}{An ecod object.}

\item{...}{Additional arguments passed to print.}
}
\description{
Print Method for ECOD Objects
}

