% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecod.R
\name{summary.ecod}
\alias{summary.ecod}
\title{Summary Method for ECOD Objects}
\usage{
\method{summary}{ecod}(object, ...)
}
\arguments{
\item{object}{An ecod object.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Summary Method for ECOD Objects
}

